fix ft functions

+++ libft/ft_putchar_fd.c
@@ -6,7 +6,7 @@
 /*   By: marielidias <marielidias@student.42.fr>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/01 12:17:31 by marielidias       #+#    #+#             */
-/*   Updated: 2024/11/01 13:33:56 by marielidias      ###   ########.fr       */
+/*   Updated: 2024/11/07 11:03:20 by marielidias      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,5 +14,5 @@
 
 void	ft_putchar_fd(char c, int fd)
 {
-	write(fd, &c, 1);
+	write((int)fd, &c, 1);
 }
